+++ new-game-project/scripts/follow_camera.gd
@@ -3,7 +3,7 @@ extends Camera3D
 
 @export_node_path("Node3D") var target_path: NodePath
 @export_group("Projection")
-@export_range(1.0, 32.0, 0.1) var orthographic_size := 12.3
+@export_range(1.0, 32.0, 0.01) var orthographic_size := 11.24
 @export_range(20.0, 120.0, 1.0) var perspective_fov := 75.0
 @export_group("Follow")
 @export_range(2.0, 30.0, 0.25) var follow_distance := 8.0
@@ -13,7 +13,7 @@ extends Camera3D
 @export_range(20.0, 89.0, 1.0) var maximum_downward_pitch := 80.0
 
 var _target: Node3D
-var _pitch := -60.0
+var _pitch := -50.0
 
 
 func _ready() -> void: